requirements.txt
Flask
stegano
Pillow